requirements.txt
colored~=1.4.2
Pillow~=8.3.2